Move account-request-authorise-consent.js to app/setup-account-request/ dir
