+++ app/setup-account-request/index.js
@@ -1,3 +1,6 @@
+const { accountRequestAuthoriseConsent, statePayload } = require('./account-request-authorise-consent');
 const { setupAccountRequest } = require('./setup-account-request');
 
+exports.accountRequestAuthoriseConsent = accountRequestAuthoriseConsent;
 exports.setupAccountRequest = setupAccountRequest;
+exports.statePayload = statePayload;
